+++ src/emitter.js
@@ -10,7 +10,7 @@ var Utils = require( './utils' );
 var BEFOREALL = 'beforeAll',
 	AFTERALL = 'afterAll'
 ;
-var specialEvents = ['immediate', BEFOREALL, AFTERALL];
+var specialEvents = [BEFOREALL, AFTERALL];
 
 // The prototype methods are stored in a different object
 // and applied as non enumerable properties later