+++ src/freezer.js
@@ -2,15 +2,16 @@
 
 var Utils = require( './utils.js' ),
 	Emitter = require( './emitter' ),
-	Mixins = require( './mixins' ),
 	Frozen = require( './frozen' )
 ;
 
 //#build
 var Freezer = function( initialValue, options ) {
 	var me = this,
-		mutable = ( options && options.mutable ) || false,
-		live = ( options && options.live ) || live
+		ops = options || {},
+		store = {
+			live: ops.live || false
+		}
 	;
 
 	// Immutable data
@@ -22,15 +23,13 @@ var Freezer = function( initialValue, options ) {
 		;
 		if( _.listener ){
 			Frozen.trigger( node, 'update', 0, true );
-
-			if( !_.parents.length )
-				_.listener.trigger('immediate', 'now');
 		}
 
 		for (i = 0; i < _.parents.length; i++) {
-			notify('now', _.parents[i]);
+			_.store.notify( 'now', _.parents[i] );
 		}
 	};
+
 	var addToPivotTriggers = function( node ){
 		pivotTriggers.push( node );
 		if( !pivotTicking ){
@@ -41,80 +40,57 @@ var Freezer = function( initialValue, options ) {
 			});
 		}
 	};
-	var notify = function notify( eventName, node, options ){
-		var _ = node.__,
-			nowNode
-		;
 
-		if( eventName == 'listener' )
-			return Frozen.createListener( node );
-
+	store.notify = function notify( eventName, node, options ){
 		if( eventName == 'now' ){
 			if( pivotTriggers.length ){
 				while( pivotTriggers.length ){
-					nowNode = pivotTriggers.shift();
-					triggerNow( nowNode );
+					triggerNow( pivotTriggers.shift() );
 				}
 			}
 			else {
 				triggerNow( node );
 			}
+
 			return node;
 		}
 
-		var update = Frozen.update( eventName, node, options );
+		var update = Frozen[eventName]( node, options );
 
 		if( eventName != 'pivot' ){
 			var pivot = Utils.findPivot( update );
 			if( pivot ) {
 				addToPivotTriggers( update );
-	  			return pivot;
+	  		return pivot;
 			}
 		}
 
 		return update;
 	};
 
-	var freeze = function(){};
-	if( !mutable )
-		freeze = function( obj ){ Object.freeze( obj ); };
+	store.freezeFn = ops.mutable === true ?
+		function(){} :
+		function( obj ){ Object.freeze( obj ); }
+	;
 
 	// Create the frozen object
-	frozen = Frozen.freeze( initialValue, notify, freeze, live );
+	frozen = Frozen.freeze( initialValue, store );
+	frozen.__.updateRoot = function( prevNode, updated ){
+		if( prevNode === frozen ){
+			frozen = updated;
+		}
+	}
 
 	// Listen to its changes immediately
-	var listener = frozen.getListener();
+	var listener = frozen.getListener(),
+		hub = {}
+	;
 
-	// Updating flag to trigger the event on nextTick
-	var updating = false;
-
-	listener.on( 'immediate', function( prevNode, updated ){
-
-		if( prevNode == 'now' ){
-			if( !updating )
-				return;
-			updating = false;
-			return me.trigger( 'update', frozen );
-		}
-
-		if( prevNode != frozen )
-			return;
-
-		frozen = updated;
-
-		if( live )
-			return me.trigger( 'update', updated );
-
-		// Trigger on next tick
-		if( !updating ){
-			updating = true;
-			Utils.nextTick( function(){
-				if( updating ){
-					updating = false;
-					me.trigger( 'update', frozen );
-				}
-			});
-		}
+	Utils.each(['on', 'off', 'once', 'trigger'], function( method ){
+		var attrs = {};
+		attrs[ method ] = listener[method].bind(listener);
+		Utils.addNE( me, attrs );
+		Utils.addNE( hub, attrs );
 	});
 
 	Utils.addNE( this, {
@@ -122,18 +98,17 @@ var Freezer = function( initialValue, options ) {
 			return frozen;
 		},
 		set: function( node ){
-			var newNode = notify( 'reset', frozen, node );
-			newNode.__.listener.trigger( 'immediate', frozen, newNode );
+			console.log('setting');
+			frozen.reset( node );
+		},
+		getEventHub: function(){
+			return hub;
 		}
 	});
 
 	Utils.addNE( this, { getData: this.get, setData: this.set } );
-
-	// The event store
-	this._events = [];
 };
 
-Freezer.prototype = Utils.createNonEnumerable({constructor: Freezer}, Emitter);
 //#build
 
 module.exports = Freezer;